+++ tests/default/twg.test.ts
@@ -1,7 +1,7 @@
 /* eslint-disable @typescript-eslint/no-unnecessary-condition */
 /* eslint-disable no-constant-binary-expression */
 
-import { twg } from "src"
+import { createTwg, twg } from "src"
 import { twg as liteTwg } from "src/lite"
 
 describe("twg()", () => {
@@ -563,65 +563,6 @@ describe("twg()", () => {
         ])('"$expected"', ({ args, expected }) => {
             expect(twg(...args)).toBe(expected)
             expect(liteTwg(...args)).toBe(expected)
-        })
-    })
-
-    describe("Custom separator:", () => {
-        it.each([
-            {
-                args: ["multiple classes", { var: "class" }],
-                expected: "multiple classes var-class"
-            }
-        ])('"$expected"', ({ args, expected }) => {
-            expect(twg(...args, { separator: "-" })).toBe(expected)
-        })
-
-        it.each([
-            {
-                args: ["multiple classes", { var: "class" }, { separator: "-" }],
-                expected: "multiple classes var-class"
-            }
-        ])('"$expected"', ({ args, expected }) => {
-            expect(twg(...args)).toBe(expected)
-        })
-
-        it.each([
-            {
-                args: ["multiple classes", { var: "class" }, { separator: "-" }],
-                expected: "multiple classes var-class separator--"
-            }
-        ])('"$expected"', ({ args, expected }) => {
-            expect(twg(...args, { separator: "-" })).toBe(expected)
-        })
-
-        it("Handles separator option in the last object:", () => {
-            expect(
-                twg(
-                    "multiple classes",
-                    {
-                        var1: "class",
-                        var2: "multiple classes"
-                    },
-                    {
-                        separator: "-"
-                    }
-                )
-            ).toBe("multiple classes var1-class var2-multiple var2-classes")
-        })
-
-        it("Handles separator option not in the last object:", () => {
-            expect(
-                twg(
-                    "multiple classes",
-                    {
-                        separator: "-"
-                    },
-                    {
-                        var1: "class",
-                        var2: "multiple classes"
-                    }
-                )
-            ).toBe("multiple classes separator:- var1:class var2:multiple var2:classes")
         })
     })