+++ tests/extend/ast.test.ts
@@ -96,7 +96,7 @@ describe("transformer()", () => {
                 expected: `twg("multiple classes", "var1:multiple var1:classes var1:var2:multiple var1:var2:classes var1:var2:multiple var1:var2:classes var1:var2:var3:multiple var1:var2:var3:classes var1:var2:var3:multiple var1:var2:var3:classes var1:var2:var3:var6:multiple var1:var2:var3:var6:classes var1:var2:var3:var7:multiple var1:var2:var3:var7:classes var1:var2:var3:var7:another var1:var2:var3:var7:class var1:var2:var3:var7:w-{{{ var1:var2:var3:var7:var9:multiple var1:var2:var3:var7:var9:classes var1:var2:var3:var7:var-10:multiple var1:var2:var3:var7:var-10:classes var1:var4:multiple var1:var4:classes var1:var4:multiple var1:var4:classes var1:var4:var5:multiple var1:var4:var5:classes var1:var4:var5:multiple var1:var4:var5:classes var1:var4:var5:another var1:var4:var5:class var1:var4:var5:var2:other var1:var4:var5:var2:multiple var1:var4:var5:var2:classes var1:var4:var5:var2:multiple var1:var4:var5:var2:classes var-4:multiple var-4:classes");`
             }
         ])('"$expected"', ({ contents, expected }) => {
-            expect(transformer(contents, { callee: "twg" })).toStrictEqual(expected)
+            expect(transformer(contents, { callee: "twg", separator: ":" })).toStrictEqual(expected)
             expect(liteTransformer(contents, "twg")).toStrictEqual(expected)
         })
     })
@@ -146,7 +146,7 @@ describe("transformer()", () => {
                 expected: `<div className={twg("multiple classes", "var1:other var1:multiple var1:classes var2:multiple var2:classes var3", className)} />;`
             }
         ])('"$expected"', ({ contents, expected }) => {
-            expect(transformer(contents, { callee: "twg" })).toStrictEqual(expected)
+            expect(transformer(contents, { callee: "twg", separator: ":" })).toStrictEqual(expected)
             expect(liteTransformer(contents, "twg")).toStrictEqual(expected)
         })
     })
@@ -182,7 +182,7 @@ describe("transformer()", () => {
                 expected: `<div className={twg("multiple classes", "var1:other var1:multiple var1:classes var2:multiple var2:classes var3", className)} />;`
             }
         ])('"$expected"', ({ contents, expected }) => {
-            expect(transformer(contents, { callee: "twg" })).toStrictEqual(expected)
+            expect(transformer(contents, { callee: "twg", separator: ":" })).toStrictEqual(expected)
             expect(liteTransformer(contents, "twg")).toStrictEqual(expected)
         })
     })
@@ -232,7 +232,7 @@ describe("transformer()", () => {
                 expected: `<div className={twg("multiple classes", "var1:other var1:multiple var1:classes var1:class var2:multiple var2:classes var2:avc var3:lorem var3:ipsum var3:class var3:other var3:multiple var3:class var3:other var3:class", className)} />;`
             }
         ])('"$expected"', ({ contents, expected }) => {
-            expect(transformer(contents, { callee: "twg" })).toStrictEqual(expected)
+            expect(transformer(contents, { callee: "twg", separator: ":" })).toStrictEqual(expected)
             expect(liteTransformer(contents, "twg")).toStrictEqual(expected)
         })
     })